Add sameSite to cookie and remove session.
Session apparently is prone to leakage. sameSite may also be a security risk.

Last crack on cross domain cookies as this is a particular pain point apparently when deploying client and server on seperate domains. Seems to be an easy fix to migrate both to Heroku, but giving it another go. Will revisit cross domain with more backend experience.